+++ src/MultiInput/MultiInput.js
@@ -199,6 +199,10 @@ MultiInput.propTypes = {
     onTagsUpdate: PropTypes.func
 };
 
+MultiInput.defaultProps = {
+    onTagsUpdate: () => {}
+};
+
 MultiInput.propDescriptions = {
     data: 'Collection of items to display in the list.',
     onTagsUpdate: 'Callback function when a tag is added or removed. Returns array of tags selected.',